pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audio-recorder-streamlit>=0.0.10",
    "gtts>=2.5.4",
    "openai-agents>=0.2.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "speechrecognition>=3.14.3",
    "streamlit>=1.50.0",
]
